+++ src/main/java/com/sna/project/phonephop/PhoneShop/service/BrandService.java
@@ -8,11 +8,11 @@ import java.util.List;
 import java.util.Map;
 
 public interface BrandService {
-    Brand findBrandById(Integer id);
+    Brand findBrandById(Long id);
     Brand SaveBrand(BrandDTO brandDTO);
 //    List<Brand> findAllBrands();
-    Brand UpdateBrandById(Integer id, BrandDTO brandDTO);
-    String DeleteBrandById(Integer id);
+    Brand UpdateBrandById(Long id, BrandDTO brandDTO);
+    String DeleteBrandById(Long id);
     List<Brand> filterByName(String name);
     List<Brand> findAllBrand(Map<String,String> param);
     Page<Brand> Pagination(Map<String,String> param);
